pyproject.toml
[project]
name = "agentmesh"
version = "0.1.0"
description = "Multi AI agent collaboration hub - unified context, memory and orchestration for Claude Code, Codex CLI and OpenClaw"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "agentmesh contributors"}]
keywords = ["ai", "agent", "orchestration", "claude", "codex", "openclaw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff", "mypy"]

[project.scripts]
agentmesh = "agentmesh.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
